[fix] ルーティングの分離
