+++ .gitignore
@@ -1,0 +1,4 @@
+# tmpを除外する（Airを採用しているため）
+*/tmp*
+
+.env
