--- public/view/schedule.html
@@ -1,65 +1,0 @@
-{{define "hello"}}
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>canvasで図形を描く</title>
-<script type="text/javascript">
-<!--
-  window.onload = function() {
-    drawCanvas();
-};
-function drawCanvas() {
-  // 描画コンテキストの取得
-  const canvas = document.getElementById('scheduleCanvas');
-  canvas.style.position = "absolute";
-  canvas.style.left = "270px";
-  canvas.style.top = "150px"
-
-  if (canvas.getContext) {
-    const context = canvas.getContext('2d');
-
-    // free-scheduleの変数定義
-    const x = 50 + (({{index . "starthour"}}-6) * 60) + {{index . "startminute"}};
-    const y = {{index . "id"}} * 70;
-    const width = (({{index . "endhour"}} - {{index . "starthour"}}) * 60) + {{index . "endminute"}} - {{index . "startminute"}};
-    const height = 70;
-    const radius = 15;
-    const lineColor = "rgba(215,155,0.1)"
-    const fillColor = "rgba(212,225,245,0.4)";
-    // const color = "rgba(212,225,245,1)";
-
-    reset();
-
-    function reset() {
-      context.clearRect(0,0,canvas.width,canvas.height);
-      drawSchedule(x, y, width, height, radius, lineColor, fillColor);
-
-    }
-
-    // スケジュールを描画する関数（角を丸める）
-    function drawSchedule(x, y, width, height, radius, lineColor, fillColor) {
-      context.beginPath();
-      context.lineWidth = 1;
-      context.strokeStyle = lineColor;
-      context.fillStyle = fillColor;
-      context.moveTo(x,y + radius);
-      context.arc(x+radius, y+height-radius, radius, Math.PI, Math.PI*0.5, true);
-      context.arc(x+width-radius, y+height-radius, radius, Math.PI*0.5, 0,1);
-      context.arc(x+width-radius, y+radius, radius, 0, Math.PI*1.5, 1);
-      context.arc(x+radius, y+radius, radius, Math.PI*1.5, Math.PI, 1);
-      context.closePath();
-      context.stroke();
-      context.fill();
-    }
-  }
-}
--->
-</script>
-</head>
-<body>
-<canvas id="scheduleCanvas" width="1225" height="670" style="background-color:rgb(210, 255, 208);">
-</canvas>
-</body>
-</html>
-{{end}}
