--- public/views/schedules.html
@@ -1,6 +1,0 @@
-{{define "schedules"}}
-{{range .}}
-{{.ID}}
-{{.EndHour}}
-{{end}}
-{{end}}
