+++ src/internal/public/views/schedules.html
@@ -1,0 +1,6 @@
+{{define "schedules"}}
+{{range .}}
+{{.ID}}
+{{.EndHour}}
+{{end}}
+{{end}}
