+++ src/mysql/Dockerfile
@@ -1,0 +1,2 @@
+FROM mysql:8.0
+ENV LANG ja_JP.UTF-8
